refactor(web): decompose Workouts.tsx and LogCard.tsx — T3 batch 3
Workouts.tsx: 592 → 214 LOC
- Extract useWorkoutsOrchestrator hook (all state + callbacks)

LogCard.tsx: 533 → 218 LOC
- Extract LogCardSearch (journal search section)
- Extract LogCardWeeklyTable (collapsible weekly macro table)
- Extract LogCardAnalytics (trend charts + top meals + meal types)

All extracted files are well under the 250 LOC target.
No circular dependencies; typecheck passes.
