+++ apps/web/src/modules/nutrition/components/LogCard.tsx
@@ -1,38 +1,19 @@
-import { useEffect, useMemo, useState } from "react";
+import { useMemo, useState } from "react";
 import type { Dispatch, SetStateAction } from "react";
-import { Card } from "@shared/components/ui/Card";
-import { SectionHeading } from "@shared/components/ui/SectionHeading";
-import { Input } from "@shared/components/ui/Input";
-import { cn } from "@shared/lib/ui/cn";
 import { ConfirmDialog } from "@shared/components/ui/ConfirmDialog";
-import { Icon } from "@shared/components/ui/Icon";
 import { EmptyState } from "@shared/components/ui/EmptyState";
-import {
-  searchMealsByName,
-  getMacrosForDateRange,
-  estimateLogBytes,
-  toLocalISODate,
-} from "../lib/nutritionStorage";
+import { estimateLogBytes, toLocalISODate } from "../lib/nutritionStorage";
 import {
   addDaysISODate,
   type Meal,
   type MealTypeId,
   type NutritionLog,
 } from "@sergeant/nutrition-domain";
-import {
-  MEAL_ORDER,
-  MEAL_META,
-  isMealTypeId,
-  mealTypeFromLabel,
-} from "../lib/mealTypes";
-import {
-  avgFromSummary,
-  getRowsForRange,
-  mealTypeBreakdown,
-  summarizeRows,
-  topMeals,
-} from "../lib/nutritionStats";
+import { isMealTypeId, mealTypeFromLabel } from "../lib/mealTypes";
 import { VirtualMealList } from "./VirtualMealList";
+import { LogCardSearch } from "./LogCardSearch";
+import { LogCardWeeklyTable } from "./LogCardWeeklyTable";
+import { LogCardAnalytics } from "./LogCardAnalytics";
 
 interface LogCardProps {
   log: NutritionLog;
@@ -84,47 +65,12 @@ export function LogCard({
   onDuplicateYesterday,
   onTrimLog,
 }: LogCardProps) {
-  const [searchQuery, setSearchQuery] = useState("");
-  const [debouncedSearchQuery, setDebouncedSearchQuery] = useState("");
   const [duplicateConfirm, setDuplicateConfirm] = useState(false);
   const [trimConfirm, setTrimConfirm] = useState(false);
-  const [statsRange, setStatsRange] = useState(30);
-  const [weekOpen, setWeekOpen] = useState(false);
-
-  useEffect(() => {
-    const id = setTimeout(() => setDebouncedSearchQuery(searchQuery), 150);
-    return () => clearTimeout(id);
-  }, [searchQuery]);
 
   const dayData = log[selectedDate];
   const meals = dayData?.meals || [];
   const groups = groupByMealType(meals);
-
-  const searchHits = useMemo(() => {
-    const q = debouncedSearchQuery.trim();
-    if (!q) return [];
-    return searchMealsByName(log, q).slice(0, 40);
-  }, [log, debouncedSearchQuery]);
-
-  const weekRows = useMemo(
-    () => getMacrosForDateRange(log, selectedDate, 7),
-    [log, selectedDate],
-  );
-
-  const statsRows = useMemo(
-    () => getRowsForRange(log, selectedDate, statsRange),
-    [log, selectedDate, statsRange],
-  );
-  const statsSummary = useMemo(() => summarizeRows(statsRows), [statsRows]);
-  const statsAvg = useMemo(() => avgFromSummary(statsSummary), [statsSummary]);
-  const statsTop = useMemo(
-    () => topMeals(log, selectedDate, statsRange, 8),
-    [log, selectedDate, statsRange],
-  );
-  const statsMealTypes = useMemo(
-    () => mealTypeBreakdown(log, selectedDate, statsRange),
-    [log, selectedDate, statsRange],
-  );
 
   const logBytes = useMemo(() => estimateLogBytes(log), [log]);
   const logSizeWarn = logBytes > 350_000;
@@ -176,99 +122,11 @@ export function LogCard({
           </button>
         )}
 
-        <Card
-          variant="flat"
-          radius="lg"
-          padding="none"
-          className="bg-panel/40 px-3 py-3 space-y-2"
-        >
-          <SectionHeading as="div" size="xs">
-            Пошук по журналу
-          </SectionHeading>
-          <Input
-            value={searchQuery}
-            onChange={(e) => setSearchQuery(e.target.value)}
-            placeholder="Назва страви…"
-            aria-label="Пошук по журналу"
-          />
-          {searchQuery.trim() && (
-            <ul className="max-h-48 overflow-y-auto text-sm space-y-1">
-              {searchHits.length === 0 && (
-                <li className="text-muted text-xs">Нічого не знайдено</li>
-              )}
-              {searchHits.map(({ date, meal }) => {
-                const mac = meal.macros || {
-                  kcal: null,
-                  protein_g: null,
-                  fat_g: null,
-                  carbs_g: null,
-                };
-                return (
-                  <li
-                    key={`${date}-${meal.id}`}
-                    className="flex items-center gap-2 bg-panelHi rounded-xl px-2.5 py-2"
-                  >
-                    <button
-                      type="button"
-                      className="text-left min-w-0 flex-1"
-                      onClick={() => {
-                        setSelectedDate(date);
-                        setSearchQuery("");
-                      }}
-                    >
-                      <div className="text-xs font-semibold text-text truncate">
-                        {meal.name}
-                      </div>
-                      <div className="flex gap-1.5 mt-0.5 flex-wrap">
-                        <span className="text-2xs text-subtle">{date}</span>
-                        {mac.kcal != null && (
-                          <span className="text-2xs text-nutrition-strong dark:text-nutrition font-bold">
-                            {Math.round(mac.kcal)} ккал
-                          </span>
-                        )}
-                        {mac.protein_g != null && (
-                          <span className="text-2xs text-subtle">
-                            Б{Math.round(mac.protein_g)}
-                          </span>
-                        )}
-                      </div>
-                    </button>
-                    <button
-                      type="button"
-                      className="shrink-0 w-8 h-8 flex items-center justify-center rounded-xl bg-nutrition/10 text-nutrition-strong dark:text-nutrition hover:bg-nutrition/20 transition-colors"
-                      onClick={() => {
-                        onAddMealFromSearch?.({
-                          id: `meal_${Date.now()}_${Math.random().toString(36).slice(2, 7)}`,
-                          time: "",
-                          name: meal.name,
-                          mealType: meal.mealType,
-                          label: meal.label,
-                          macros: meal.macros
-                            ? { ...meal.macros }
-                            : {
-                                kcal: null,
-                                protein_g: null,
-                                fat_g: null,
-                                carbs_g: null,
-                              },
-                          source: "manual",
-                          macroSource: "manual",
-                          foodId: null,
-                          amount_g: null,
-                        });
-                        setSearchQuery("");
-                      }}
-                      title="Додати до поточного дня"
-                      aria-label={`Додати ${meal.name} до поточного дня`}
-                    >
-                      +
-                    </button>
-                  </li>
-                );
-              })}
-            </ul>
-          )}
-        </Card>
+        <LogCardSearch
+          log={log}
+          setSelectedDate={setSelectedDate}
+          onAddMealFromSearch={onAddMealFromSearch}
+        />
 
         {logSizeWarn && (
           <div className="rounded-2xl border border-amber-500/40 bg-amber-500/10 px-3 py-2 text-xs text-amber-100">
@@ -283,184 +141,9 @@ export function LogCard({
           </div>
         )}
 
-        <SectionHeading
-          as="button"
-          size="xs"
-          type="button"
-          onClick={() => setWeekOpen((v) => !v)}
-          className="flex items-center gap-2 w-full text-left py-1"
-        >
-          <Icon
-            name="chevron-right"
-            size={12}
-            strokeWidth={2.5}
-            className={cn(
-              "transition-transform shrink-0",
-              weekOpen ? "rotate-90" : "",
-            )}
-          />
-          Журнал за тиждень
-        </SectionHeading>
+        <LogCardWeeklyTable log={log} selectedDate={selectedDate} />
 
-        {weekOpen && (
-          <div className="rounded-2xl border border-line bg-panel/40 px-3 py-3">
-            <div className="overflow-x-auto">
-              <table className="w-full text-xs text-left">
-                <thead>
-                  <tr className="text-subtle">
-                    <th className="py-1 pr-2">Дата</th>
-                    <th className="py-1 pr-2">Ккал</th>
-                    <th className="py-1 pr-2">Б</th>
-                    <th className="py-1 pr-2">Ж</th>
-                    <th className="py-1">В</th>
-                  </tr>
-                </thead>
-                <tbody>
-                  {weekRows.map((r) => (
-                    <tr key={r.date} className="border-t border-line/40">
-                      <td className="py-1 pr-2 font-mono text-2xs">
-                        {r.date.slice(5)}
-                      </td>
-                      <td className="py-1 pr-2">{Math.round(r.kcal)}</td>
-                      <td className="py-1 pr-2">{Math.round(r.protein_g)}</td>
-                      <td className="py-1 pr-2">{Math.round(r.fat_g)}</td>
-                      <td className="py-1">{Math.round(r.carbs_g)}</td>
-                    </tr>
-                  ))}
-                </tbody>
-              </table>
-            </div>
-          </div>
-        )}
-
-        <div className="rounded-2xl border border-line bg-panel/40 px-3 py-3 space-y-3">
-          <div className="flex items-center justify-between gap-2">
-            <SectionHeading as="div" size="xs">
-              Аналітика (тренди)
-            </SectionHeading>
-            <div className="flex gap-2">
-              {[30, 90].map((d) => (
-                <button
-                  key={d}
-                  type="button"
-                  onClick={() => setStatsRange(d)}
-                  className={cn(
-                    "px-2 py-1 rounded-xl text-xs font-semibold border",
-                    statsRange === d
-                      ? "border-nutrition/60 text-nutrition-strong dark:text-nutrition bg-nutrition/10"
-                      : "border-line text-subtle bg-panelHi",
-                  )}
-                >
-                  {d} днів
-                </button>
-              ))}
-            </div>
-          </div>
-
-          <div className="grid grid-cols-2 sm:grid-cols-4 gap-2">
-            {[
-              { key: "kcal", label: "Сер. ккал/день", v: statsAvg.kcal },
-              { key: "protein_g", label: "Сер. Б/день", v: statsAvg.protein_g },
-              { key: "fat_g", label: "Сер. Ж/день", v: statsAvg.fat_g },
-              { key: "carbs_g", label: "Сер. В/день", v: statsAvg.carbs_g },
-            ].map((x) => (
-              <div key={x.key} className="bg-panelHi rounded-2xl px-2 py-3">
-                <div className="text-2xs text-subtle">{x.label}</div>
-                <div className="text-base font-extrabold text-text tabular-nums">
-                  {Math.round(Number(x.v) || 0)}
-                </div>
-                <div className="text-2xs text-subtle">
-                  на {statsAvg.denom} активн. днів
-                </div>
-              </div>
-            ))}
-          </div>
-
-          <div className="bg-panelHi rounded-2xl px-3 py-3">
-            <SectionHeading as="div" size="xs" className="mb-2">
-              Калорії по днях (останні {Math.min(statsRange, statsRows.length)})
-            </SectionHeading>
-            {statsRows.length === 0 ? (
-              // eslint-disable-next-line sergeant-design/no-bare-empty-text -- pre-existing tech debt; tracked in docs/tech-debt/frontend.md
-              <div className="text-xs text-muted">Поки що порожньо</div>
-            ) : (
-              (() => {
-                const kcals = statsRows.map((r) => Number(r.kcal) || 0);
-                const max = Math.max(1, ...kcals);
-                return (
-                  <div className="flex items-end gap-0.5 h-12">
-                    {kcals.slice(-statsRange).map((k, i) => (
-                      <div
-                        key={i}
-                        title={`${Math.round(k)} ккал`}
-                        className="flex-1 rounded-sm bg-nutrition/60"
-                        style={{
-                          height: `${Math.max(2, Math.round((k / max) * 48))}px`,
-                        }}
-                      />
-                    ))}
-                  </div>
-                );
-              })()
-            )}
-          </div>
-
-          <div className="grid grid-cols-1 sm:grid-cols-2 gap-2">
-            <div className="bg-panelHi rounded-2xl px-3 py-3">
-              <SectionHeading as="div" size="xs" className="mb-2">
-                Топ страв
-              </SectionHeading>
-              {statsTop.length === 0 ? (
-                // eslint-disable-next-line sergeant-design/no-bare-empty-text -- pre-existing tech debt; tracked in docs/tech-debt/frontend.md
-                <div className="text-xs text-muted">Поки що порожньо</div>
-              ) : (
-                <ol className="space-y-1">
-                  {statsTop.map((x) => (
-                    <li
-                      key={x.name}
-                      className="flex items-baseline justify-between gap-2"
-                    >
-                      <span className="text-xs text-text truncate">
-                        {x.name}
-                      </span>
-                      <span className="text-xs text-subtle shrink-0">
-                        {x.count}× · {Math.round(x.kcal)} ккал
-                      </span>
-                    </li>
-                  ))}
-                </ol>
-              )}
-            </div>
-            <div className="bg-panelHi rounded-2xl px-3 py-3">
-              <SectionHeading as="div" size="xs" className="mb-2">
-                Розподіл прийомів
-              </SectionHeading>
-              {Object.keys(statsMealTypes).length === 0 ? (
-                // eslint-disable-next-line sergeant-design/no-bare-empty-text -- pre-existing tech debt; tracked in docs/tech-debt/frontend.md
-                <div className="text-xs text-muted">Поки що порожньо</div>
-              ) : (
-                <ul className="space-y-1">
-                  {MEAL_ORDER.filter(
-                    (t) => (statsMealTypes[t]?.count ?? 0) > 0,
-                  ).map((t) => (
-                    <li
-                      key={t}
-                      className="flex items-baseline justify-between gap-2"
-                    >
-                      <span className="text-xs text-text">
-                        {MEAL_META[t]?.emoji} {MEAL_META[t]?.label || t}
-                      </span>
-                      <span className="text-xs text-subtle shrink-0">
-                        {statsMealTypes[t]!.count}× ·{" "}
-                        {Math.round(statsMealTypes[t]!.kcal)} ккал
-                      </span>
-                    </li>
-                  ))}
-                </ul>
-              )}
-            </div>
-          </div>
-        </div>
+        <LogCardAnalytics log={log} selectedDate={selectedDate} />
 
         {meals.length === 0 ? (
           <EmptyState
